requirements.txt
flask
mysql-connector-python
xhtml2pdf
